Migrate from requirements.txt to Poetry
`requirements.txt` has a few issues with regards to proper dependency
locking and resolution (frequently barely doing one and not even doing
the other). Poetry is a packaging and dependency manager which has a
proper resolution algorithm that can lock versions of packages and
create an accurate dependency tree which will have deterministic and
repeatable installs and, thus, builds.